feat: create location endpoint

+++ Core/Entities/Location.cs
@@ -1,2 +1,2 @@
 namespace Core.Entities;
-public record Location(Guid Id, string Name, int OpeningTimeInSeconds, int ClosingTimeInSeconds);
+public record Location(Guid Id, string Name, long OpeningTimeInSeconds, long ClosingTimeInSeconds);
